♻️refactor: CodeRabbitレビュー指摘を対応（ルートパス定数化・型修正・ConsumerWidget化・エラーログ等）
- route_path_constants: kDeliverySegment等のセグメント定数を追加しパスを合成に変更（他モジュールと整合）
- delivery_quiz_config: itemPriceYen を double → int に修正（整数値に適切な型を使用）
- delivery_mock_data: buildSync() 呼び出しをトップレベル定数にキャッシュ（呼び出しのたびに再生成を防ぐ）
- delivery_app_state: コンストラクタに assert 追加、totalPrice の不要な toInt() を削除
- delivery_app_scaffold: ref.select でviewStateのみ監視、_BrowsingView/_TrackingView を ConsumerWidget 化、BuildContext フィールドのアンチパターンを除去（不要な全体再ビルドを防ぐ）
- 全4クイズ Notifier: catch (_) {} → catch (e, st) + appLogger.e() でエラーを記録
- テスト2件: Future.delayed → Completer パターンに変更（フレーキーなタイミング依存を解消）

+++ packages/quizzes/delivery/lib/src/domain/delivery_quiz_config.dart
@@ -11,7 +11,7 @@ abstract final class DeliveryQuizConfig {
   static const double trackingClearExtent = 0.7;
 
   /// 商品単価（円）
-  static const double itemPriceYen = 890;
+  static const int itemPriceYen = 890;
 
   /// ピザカテゴリのID（Quiz1のクリア判定に使用）
   static const String pizzaCategoryId = 'pizza';
